8th

+++ lib/libft/src/ft_strcmp.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-size_t	ft_strcmp(const char *s1, const char *s2)
+int	ft_strcmp(const char *s1, const char *s2)
 {
 	unsigned char *ptr1;
 	unsigned char *ptr2;
